requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-session==0.6.0
python-dotenv==1.0.1
pytest==8.0.2
pytest-flask==1.3.0
pytest-cov==4.1.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
sentry-sdk[flask]==1.40.5
psycopg2-binary==2.9.9 